third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/worker-src/service-self.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="worker-src 'self'">
<script>
  var url = new URL("/security/contentSecurityPolicy/resources/ping.js", document.baseURI).toString();
  assert_service_worker_is_loaded(url, "Same-origin service worker allowed by 'self'.");
</script>
